+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-## UPCOMING
+## 2.0.0 (UPCOMING)
+
+* Changed the public interface. See [#332](https://github.com/savonrb/savon/issues/332) and read
+  the [documentation](http://savonrb.com/version2.html) to get to know what changed and why.
 
 * Update: Use Nori 2.0 to minimalize global state. All Nori options are now encapsulated and can
   be configured through Savon.